Fix: Login logout bug causing Register Class to Fail
