+++ app/src/main/java/com/vastgk/paytap/Dashboard.java
@@ -3,7 +3,9 @@ package com.vastgk.paytap;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.Context;
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.MenuItem;
@@ -17,14 +19,19 @@ import com.android.volley.RequestQueue;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
+import com.mikhaellopez.circularimageview.CircularImageView;
 
 import org.json.JSONException;
 import org.json.JSONObject;
 
 import de.hdodenhof.circleimageview.CircleImageView;
 
+import static com.vastgk.paytap.OTPVerify.IS_LOGGED_IN;
+import static com.vastgk.paytap.OTPVerify.PREFERENCE_FILE_KEY;
+import static com.vastgk.paytap.OTPVerify.USERID;
+
 public class Dashboard extends AppCompatActivity {
-    CircleImageView profileImgView;
+    CircularImageView profileImgView;
     TextView tv_fullTransaction;
     BottomNavigationView bottomNavigationView;
 
@@ -32,11 +39,11 @@ public class Dashboard extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_dashboard);
-
+        SaveUserState();
         bottomNavigationView = findViewById(R.id.bottom_navigation);
-        profileImgView=findViewById(R.id.ProfileImgView);
-        profileImgView.setOnClickListener((v )->{
-
+        profileImgView = findViewById(R.id.ProfileImgView);
+        profileImgView.setOnClickListener((v) -> {
+            logout();
 
         });
 
@@ -44,10 +51,10 @@ public class Dashboard extends AppCompatActivity {
 
         bottomNavigationView.setOnNavigationItemSelectedListener(new BottomNavigationView.OnNavigationItemSelectedListener() {
             Intent intent;
+
             @Override
             public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-                switch (item.getItemId())
-                {
+                switch (item.getItemId()) {
                     case R.id.bottom_navigation_Dashboard_nfc:
                         Toast.makeText(Dashboard.this, "NfC", Toast.LENGTH_SHORT).show();
                         break;
@@ -64,57 +71,26 @@ public class Dashboard extends AppCompatActivity {
 
     }
 
-    public static class Registration extends AppCompatActivity {
-
-
-        Button btn_act_registration;
-        EditText mobileNumberEditText, Name_editTxt;
-
-        @Override
-        protected void onCreate(Bundle savedInstanceState) {
-            super.onCreate(savedInstanceState);
-            setContentView(R.layout.activity_registration);
-            mobileNumberEditText = findViewById(R.id.OTPVERIFY_mobileEditTxt);
-            Name_editTxt = findViewById(R.id.OTPVERIFY_ediText_Name);
-            RequestQueue requestQueue = Volley.newRequestQueue(this);
-            mobileNumberEditText.setText(getIntent().getStringExtra("mobilenumber").toString());
-            btn_act_registration = findViewById(R.id.btn_act_registration);
-            Name_editTxt.requestFocus();
-            btn_act_registration.setOnClickListener((view) -> {
-                if (Name_editTxt.getText().toString().isEmpty()) {
-                    Name_editTxt.setError("Enter The Name First");
-                    return;
-
-                }
-                String url = "http://api.nixbymedia.com/paytap/users_add.php?username=" + mobileNumberEditText.getText() + "&name=" + Name_editTxt.getText();
-                StringRequest request = new StringRequest(Request.Method.GET, url, response -> {
-                    try {
-                        JSONObject jo = new JSONObject(response);
-                        Toast.makeText(this, "Registering The user Now " + response, Toast.LENGTH_SHORT).show();
-                        //user is added Successfully and now goto OTP verify Class
-                        Intent intent = new Intent(Registration.this, OTPVerify.class);
-                        intent.putExtra("mobilenumber", getIntent().getStringExtra("mobilenumber"));
-                        intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_NEW_TASK);
-                        startActivity(intent);
-                        finish();
-
-
-                    } catch (JSONException e) {
-                        e.printStackTrace();
-                        Log.e("REGSITRATOIN", "onCreate: " + e.getLocalizedMessage());
-                    }
-
-
-                }, error -> {
-
-                    Log.e("REGISTRATION", "onCreate:" + error.getLocalizedMessage());
-                });
-
-                requestQueue.add(request);
-
-
-            });
-
-        }
+    private void SaveUserState() {
+        SharedPreferences sharedPreferences=getSharedPreferences( PREFERENCE_FILE_KEY, Context.MODE_PRIVATE);
+        SharedPreferences.Editor editor=sharedPreferences.edit();
+        editor.putBoolean(IS_LOGGED_IN,true);
+        editor.putString(USERID,getIntent().getStringExtra("mobilenumber"));
+        editor.commit();
     }
+
+    private void logout() {
+        SharedPreferences sharedPreferences = getSharedPreferences(PREFERENCE_FILE_KEY, MODE_PRIVATE);
+        SharedPreferences.Editor editor = sharedPreferences.edit();
+        editor.putBoolean(IS_LOGGED_IN, false);
+        editor.putString(USERID, "null");
+        editor.commit();
+        Intent intent = new Intent(Dashboard.this, OTPVerification.class);
+        intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
+        startActivity(intent);
+        finish();
+
+    }
+
+
 }
