+++ app/src/main/java/com/vastgk/paytap/MainActivity.java
@@ -7,10 +7,13 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.os.Handler;
+import android.util.Log;
 
 import com.google.firebase.auth.FirebaseAuth;
 
 public class MainActivity extends AppCompatActivity {
+
+    private static final String TAG = "MAINDEBUG";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -18,7 +21,7 @@ public class MainActivity extends AppCompatActivity {
         setContentView(R.layout.activity_main);
         SharedPreferences sharedPreferences=MainActivity.this.getSharedPreferences(OTPVerify.PREFERENCE_FILE_KEY, Context.MODE_PRIVATE);
         boolean isLoggedin=sharedPreferences.getBoolean(OTPVerify.IS_LOGGED_IN,false);
-
+        Log.d(TAG, "onCreate: "+ (isLoggedin ?"true":"false"));
         new Handler().postDelayed(() -> {
 
                     if (!isLoggedin) {
@@ -29,7 +32,7 @@ public class MainActivity extends AppCompatActivity {
                         finish();
                     }else
                     {
-                        Intent intent = new Intent(MainActivity.this, OTPVerification.class);
+                        Intent intent = new Intent(MainActivity.this, Dashboard.class);
                         intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK|Intent.FLAG_ACTIVITY_NEW_TASK);
                         intent.putExtra("userid",sharedPreferences.getString(OTPVerify.USERID,"NO_MOBILENUMBER_IN_SHAREDPREFERENCE"));
                         MainActivity.this.startActivity(intent);
